+++ ChangeLog.md
@@ -9,6 +9,8 @@ Blob:
 - Fixed issue of: blob batch subresponse is slightly different from the on from Azure serivce, which causes exception in CPP SDK.
 - Fixed issue of: setMetadata API allows invalid metadata name with hyphen.
 - Supported rest API GetBlobTag, SetBlobTag.
+- Supported set Blob Tags in upload blob, copy blob.
+- Supported get Blob Tags (count) in download blob, get blob properties, list blobs.
 
 ## 2023.03 Version 3.23.0